requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
psycopg2-binary
psycopg[binary]
Pillow
gunicorn
dj-database-url
whitenoise
python-dotenv
asgiref
sqlparse
tzdata
